- fixed %preun scripts, - release++.
Changed files:
    xpdf-chinese-traditional.spec -> 1.9

+++ xpdf-chinese-traditional.spec
@@ -2,7 +2,7 @@ Summary:	Big5 and Big5ascii encoding support for xpdf
 Summary(pl):	Wsparcie kodowania Big5 i Big5ascii dla xpdf
 Name:		xpdf-chinese-traditional
 Version:	1.0
-Release:	3
+Release:	4
 License:	GPL
 Group:		X11/Applications
 Source0:	ftp://ftp.foolabs.com/pub/xpdf/%{name}.tar.gz
@@ -76,14 +76,16 @@ else
 fi
 
 %preun
-umask 022
-grep -v 'Big5\.unicodeMap' /etc/xpdfrc > /etc/xpdfrc.new
-grep -v 'Big5ascii\.unicodeMap' /etc/xpdfrc.new > /etc/xpdfrc
-grep -v 'Adobe-CNS1\.cidToUnicode' /etc/xpdfrc > /etc/xpdfrc.new
-grep -v 'CMap-chinese-traditional' /etc/xpdfrc.new > /etc/xpdfrc
-grep -v -e '-\*-fixed-medium-r-normal-\*-%s-\*-\*-\*-\*-\*-big5-0' /etc/xpdfrc > /etc/xpdfrc.new
-grep -v -e '-arphic-ar pl kaitim big5-medium-r-normal--%s-\*-\*-\*-c-\*-iso10646-1' /etc/xpdfrc.new > /etc/xpdfrc
-rm -f /etc/xpdfrc.new
+if [ "$1" = "0" ]; then
+	umask 022
+	grep -v 'Big5\.unicodeMap' /etc/xpdfrc > /etc/xpdfrc.new
+	grep -v 'Big5ascii\.unicodeMap' /etc/xpdfrc.new > /etc/xpdfrc
+	grep -v 'Adobe-CNS1\.cidToUnicode' /etc/xpdfrc > /etc/xpdfrc.new
+	grep -v 'CMap-chinese-traditional' /etc/xpdfrc.new > /etc/xpdfrc
+	grep -v -e '-\*-fixed-medium-r-normal-\*-%s-\*-\*-\*-\*-\*-big5-0' /etc/xpdfrc > /etc/xpdfrc.new
+	grep -v -e '-arphic-ar pl kaitim big5-medium-r-normal--%s-\*-\*-\*-c-\*-iso10646-1' /etc/xpdfrc.new > /etc/xpdfrc
+	rm -f /etc/xpdfrc.new
+fi
 
 %files
 %defattr(644,root,root,755)
